imagegen/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
diffusers>=0.27.0
transformers>=4.40.0
torch>=2.2.0
accelerate>=0.29.0
Pillow>=10.0.0
huggingface-hub>=0.22.0
